chore:

+++ start.sh
@@ -1,6 +1,14 @@
 
-echo "删除packages下除ui之外的其他目录"
-rm -rf packages/!(ui)
+#
+echo "即将删除packages下除ui之外的其他目录，按y确认"
+read -p "y/n: " confirm
+if [ $confirm != "y" ]; then
+  echo "取消操作"
+  exit 1
+fi
+find ./packages -mindepth 1 -maxdepth 1 -type d ! -name 'ui' -exec rm -rf {} +
+
+echo "删除成功"
 
 echo "安装pnpm 8.15.7, 前提是已经安装了nodejs"
 npm install -g pnpm@8.15.7
@@ -19,5 +27,3 @@ npm run build
 echo "构建完成"
 echo "启动服务"
 npm run start
-
-
